+++ lib/slack/logger.rb
@@ -1,0 +1,9 @@
+require 'logger'
+
+module Slack
+  class Logger < ::Logger
+    def self.default
+      new STDOUT
+    end
+  end
+end