+++ lib/slack/real_time/event_handlers/pref_change.rb
@@ -6,6 +6,7 @@ module Slack
         # @see https://api.slack.com/events/pref_change
         # @see https://github.com/dblock/slack-api-ref/blob/master/events/pref_change.json
         def self.call(client, data)
+          client.self.prefs ||= {}
           client.self.prefs[data.name] = data.value
         end
       end