+++ lib/slack/messages/message.rb
@@ -4,6 +4,12 @@ module Slack
       def presence
         super['presence']
       end
+
+      def to_s
+        keys.sort_by(&:to_s).map do |key|
+          "#{key}=#{self[key]}"
+        end.join(', ')
+      end
     end
   end
 end