+++ spec/slack/real_time/concurrency/eventmachine_spec.rb
@@ -4,7 +4,8 @@ begin
   RSpec.describe Slack::RealTime::Concurrency::Eventmachine::Socket do
     context 'with url' do
       let(:url) { 'wss://ms174.slack-msgs.com/websocket/xyz' }
-      let(:socket) { described_class.new(url, ping: 42) }
+      let(:logger) { ::Logger.new($stdout) }
+      let(:socket) { described_class.new(url, ping: 42, logger: logger) }
       let(:ws) { double(Faye::WebSocket::Client) }
       describe '#initialize' do
         it 'sets url' do
@@ -21,7 +22,7 @@ begin
           expect(socket.instance_variable_get('@driver')).to eq ws
         end
         it 'pings every 30s' do
-          expect(Faye::WebSocket::Client).to receive(:new).with(url, nil, ping: 42).and_return(ws)
+          expect(Faye::WebSocket::Client).to receive(:new).with(url, nil, ping: 42, logger: logger).and_return(ws)
           socket.connect!
         end
       end