+++ spec/support/real_time/connected_client.rb
@@ -7,7 +7,7 @@ RSpec.shared_context 'connected client' do |opts|
     Slack::RealTime.configure do |config|
       config.concurrency = Slack::RealTime::Concurrency::Mock
     end
-    allow(Slack::RealTime::Socket).to receive(:new).with(url, ping: 30).and_return(socket)
+    allow(Slack::RealTime::Socket).to receive(:new).with(url, ping: 30, logger: Slack::Logger.default).and_return(socket)
     allow(socket).to receive(:start_sync).and_yield
     allow(socket).to receive(:connect!).and_yield(ws)
     allow(ws).to receive(:on)